refact: move types from model.ts to types.ts
